pyproject.toml
[tool.poetry]
name = "miyoka"
version = "0.1.0"
description = ""
authors = ["FGC Replay Miyoka <fgcreplaymiyoka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-storage = "^2.16.0"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"
pygetwindow = "^0.0.9"
python-dotenv = "^1.0.1"
google-cloud-vision = "^3.7.2"
python-json-logger = "^2.0.7"
black = "^24.4.2"
flake8 = "^7.0.0"
dependency-injector = {extras = ["yaml"], version = "^4.41.0"}
google-cloud-bigquery = "^3.22.0"
google-cloud-run = "^0.10.5"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.2"
pandas = "^2.2.2"
db-dtypes = "^1.3.0"

[tool.poetry.group.win]
optional = true

[tool.poetry.group.win.dependencies]
dxcam = "^0.0.5"
pydirectinput = "^1.0.4"

[tool.poetry.group.frontend]
optional = true

[tool.poetry.group.frontend.dependencies]
streamlit = "^1.37.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
